Split ConferenceController into Conference/{Index,Show}ConferenceController
