+++ src/Controller/Conference/ShowConferenceController.php
@@ -2,14 +2,13 @@
 
 declare(strict_types=1);
 
-namespace App\Controller;
+namespace App\Controller\Conference;
 
 use App\Entity\Comment;
 use App\Entity\Conference;
 use App\Form\CommentFormType;
 use App\Message\CommentMessage;
 use App\Repository\CommentRepository;
-use App\Repository\ConferenceRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\FormInterface;
@@ -24,7 +23,7 @@ use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Twig\Environment;
 
-final class ConferenceController extends AbstractController
+final class ShowConferenceController extends AbstractController
 {
     private Environment $twig;
     private EntityManagerInterface $entityManager;
@@ -38,40 +37,6 @@ final class ConferenceController extends AbstractController
         $this->twig = $twig;
         $this->entityManager = $entityManager;
         $this->messageBus = $messageBus;
-    }
-
-    /**
-     * @Route("/")
-     */
-    public function indexNoLocale(): Response
-    {
-        return $this->redirectToRoute('homepage', ['_locale' => 'en']);
-    }
-
-    /**
-     * @Route("/{_locale<%app.supported_locales%>}/", name="homepage")
-     */
-    public function index(ConferenceRepository $conferenceRepository): Response
-    {
-        $response = new Response($this->twig->render('conference/index.html.twig', [
-            'conferences' => $conferenceRepository->findAll(),
-        ]));
-        $response->setSharedMaxAge(3600);
-
-        return $response;
-    }
-
-    /**
-     * @Route("/{_locale<%app.supported_locales%>}/conference_header", name="conference_header")
-     */
-    public function conferenceHeader(ConferenceRepository $conferenceRepository): Response
-    {
-        $response = new Response($this->twig->render('conference/header.html.twig', [
-            'conferences' => $conferenceRepository->findAll(),
-        ]));
-        $response->setSharedMaxAge(3600);
-
-        return $response;
     }
 
     /**
